backend_Pitch/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiofiles>=23.0.0
pydantic>=2.0.0,<3.0.0
email-validator>=2.0.0

# Local Speech-to-Text with Whisper (fallback)
openai-whisper
torch
torchaudio

# Real-time Speech-to-Text with Vosk
vosk>=0.3.45

# Machine Learning for Voice Analysis
transformers>=4.21.0
scikit-learn>=1.3.0
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
vaderSentiment>=3.3.2
textstat>=0.7.3
spacy>=3.7.0

# Phase 4: Advanced Features - Video and Emotion Analysis
opencv-python>=4.8.0
deepface>=0.0.79
mediapipe==0.10.14
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
kaleido>=0.2.1
scipy>=1.11.0
pandas>=2.0.0

# Enhanced Voice Emotion Analysis
tensorflow>=2.13.0
keras>=2.13.0

# Audio processing and format conversion
pydub>=0.25.1
